requirements.txt
uvicorn==0.29.0
fastapi==0.110.0
requests==2.31.0
SQLAlchemy==2.0.28
psycopg2-binary==2.9.9
python-decouple==3.8


# Testes
pytest==8.1.1

# Tools
alembic==1.13.1
httpx==0.27.0